+++ kairos-server/src/routes/deposit.rs
@@ -6,15 +6,9 @@ use axum_extra::routing::TypedPath;
 use tracing::*;
 
 use kairos_tx::asn::{SigningPayload, TransactionBody};
+use kairos_types::transactions::{Signed, Transaction};
 
-use crate::{
-    routes::PayloadBody,
-    state::{
-        transactions::{Signed, Transaction},
-        BatchStateManager,
-    },
-    AppErr,
-};
+use crate::{routes::PayloadBody, state::BatchStateManager, AppErr};
 
 #[derive(TypedPath, Debug, Clone, Copy)]
 #[typed_path("/api/v1/deposit")]
