+++ kairos-server/src/routes/withdraw.rs
@@ -6,9 +6,9 @@ use axum_extra::routing::TypedPath;
 use tracing::*;
 
 use kairos_tx::asn::{SigningPayload, TransactionBody};
+use kairos_types::transactions::{Signed, Transaction, Withdraw};
 
 use crate::routes::PayloadBody;
-use crate::state::transactions::{Signed, Transaction, Withdraw};
 use crate::state::BatchStateManager;
 use crate::AppErr;
