+++ kairos-types/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "kairos-types"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[features]
+default = ["std"]
+std = []
+
+[dependencies]
+kairos-tx = { path = "../kairos-tx" }
