+++ kairos-types/src/lib.rs
@@ -1,0 +1,5 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+extern crate alloc;
+
+pub mod transactions;
